fix processing event log dirs
